+++ playground/games/account/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rpgjs/playground-account
 
+## 1.0.0-rc.6
+
+### Patch Changes
+
+- Updated dependencies [d88267a]
+- Updated dependencies
+  - @rpgjs/common@5.0.0-rc.5
+  - @rpgjs/server@5.0.0-rc.6
+  - @rpgjs/client@5.0.0-rc.6
+  - @rpgjs/account@5.0.0-rc.6
+  - @rpgjs/vite@5.0.0-rc.6
+
 ## 1.0.0-rc.5
 
 ### Patch Changes